src/requirements.downloader.txt
requests
rq
psycopg2
openai-whisper